move migraiton target type enum to utils
