+++ extensions/sql-migration/src/models/stateMachine.ts
@@ -13,7 +13,7 @@ import * as constants from '../constants/strings';
 import * as nls from 'vscode-nls';
 import { v4 as uuidv4 } from 'uuid';
 import { sendSqlMigrationActionEvent, TelemetryAction, TelemetryViews, logError } from '../telemetry';
-import { hashString, deepClone, getBlobContainerNameWithFolder, Blob, getLastBackupFileNameWithoutFolder } from '../api/utils';
+import { hashString, deepClone, getBlobContainerNameWithFolder, Blob, getLastBackupFileNameWithoutFolder, MigrationTargetType } from '../api/utils';
 import { SKURecommendationPage } from '../wizard/skuRecommendationPage';
 import { excludeDatabases, getEncryptConnectionValue, getSourceConnectionId, getSourceConnectionProfile, getSourceConnectionServerInfo, getSourceConnectionString, getSourceConnectionUri, getTrustServerCertificateValue, SourceDatabaseInfo, TargetDatabaseInfo } from '../api/sqlUtils';
 import { LoginMigrationModel } from './loginMigrationModel';
@@ -55,12 +55,6 @@ export enum State {
 export enum ServiceTier {
 	GeneralPurpose = 'GeneralPurpose',
 	BusinessCritical = 'BusinessCritical',
-}
-
-export enum MigrationTargetType {
-	SQLVM = 'AzureSqlVirtualMachine',
-	SQLMI = 'AzureSqlManagedInstance',
-	SQLDB = 'AzureSqlDatabase'
 }
 
 export enum MigrationSourceAuthenticationType {
